Split eyefultower download into a separate file
