+++ nerfstudio/scripts/downloads/utils.py
@@ -1,0 +1,32 @@
+# Copyright 2022 the Regents of the University of California, Nerfstudio Team and contributors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+from pathlib import Path
+
+from nerfstudio.configs.base_config import PrintableConfig
+
+
+@dataclass
+class DatasetDownload(PrintableConfig):
+    """Download a dataset"""
+
+    capture_name = None
+
+    save_dir: Path = Path("data/")
+    """The directory to save the dataset to"""
+
+    def download(self, save_dir: Path) -> None:
+        """Download the dataset"""
+        raise NotImplementedError
